Mojo: Fix another sync call on associated interface hang.
For some amount of time after a sync IPC has failed because its receiver
peer has closed, MultiplexRouter is in a state where it has marked a
peer as being closed, but the task to update the AssociatedRemote of
this information hasn't been ran (or posted) yet. During this time
any new sync calls would hang indefinitely. This CL fixes this by not
waiting if a endpoint is in this state.

Bug: 436965298
Change-Id: Ibb6711e79152fd83819ad094098af6187e2d4130
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6825792
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1498604}