[iOS] Fix button to show/hide Reader mode from AI hub
This CL fixes the button to show/hide Reader mode from the AI hub. This
was broken because the action was performed asynchronously at the time
when the sheet is dismissed, but at this point `weakSelf` would be `nil`
so nothing would happen. Instead, a `weakReaderModeHandler` is now used
which would still be alive when the sheet is dismissed. A test is added
in the follow-up CL to prevent future regression.

Bug: 435413405
Change-Id: I6a8eba7691089312b7b8af6a3a20e538af7dfbb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6818489
Auto-Submit: Quentin Pubert <qpubert@google.com>
Reviewed-by: Adam Arcaro <adamta@google.com>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Cr-Commit-Position: refs/heads/main@{#1498642}