[Editing] Fix input event dataTransfer property for text drop operations
Currently, the input event fired for drop operations (insertFromDrop)
has a null dataTransfer property, which violates the specification [1]
that requires dataTransfer to be available for certain input types. The
implementation is a follow-up CL for [2]. The DataTransfer is only
attached for contenteditable elements, excluding form controls (textarea
and input elements) to maintain existing behavior and comply with the
specification. The behavior is working as expected on Safari and Firefox.
The Chrome Status entry [3] has also been created for I2S. Same Bug ID
for tracking has been used for insertFromDrop event also as
insertFromPaste.

[1] - https://w3c.github.io/input-events/#dom-inputevent-datatransfer

[2] - https://chromium-review.googlesource.com/c/chromium/src/+/6687446

[3] - https://chromestatus.com/feature/6715253274181632

Bug: 401593412
Change-Id: I30e8f2009a89b272f6712ed892e2e46e4984232e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6817846
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Pranav Modi <pranavmodi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1498652}