[UNO] Update sync toggle's visibility on route changed
Until now, the sync toggles' visibility would be updated when the
component `sync-controls` was loaded. However, this was happening while
the current route was not the account setting page, causing the toggles
to be hidden until a refresh. This changes the computation of the
visibility to happen upon navigation.

This change is behind the feature flag
`ReplaceSyncPromosWithSignInPromos`.

Toggles in question: https://screenshot.googleplex.com/9tVuTSoVXZtbW5k

Fixed: 433895051
Change-Id: Ibeb059fa9a68c0fd771f6b4198b544a320b4c944
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6823676
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Amelie Schneider <amelies@google.com>
Cr-Commit-Position: refs/heads/main@{#1498656}