Improve tracking of OAC state in SiteInfo
This CL makes more explicit what the OAC status for the SiteInfo was
(origin-keyed or site-keyed, whether due to the presence of an OAC header or by
default). This is a preliminary CL for replacing
SiteInfo::requests_origin_keyed_process by checking whether the AgentClusterKey
in SiteInfo is origin-keyed. This is because currently
requests_origin_keyed_process is only set due to OAC, while the AgentClusterKey
can be origin-keyed for other reasons. Therefore, we should track precisely
what the OAC status was on top of the AgentClusterKey.

The CL also refactors the tracking of OAC state and requests, so that it is
clearer what isolation is requested and granted. There is no behavior change in
this CL.

Bug: 342365078
Change-Id: Ia4701142b205bc3bae89f6b056eed4da8db48d10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6802381
Commit-Queue: Camille Lamy <clamy@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Auto-Submit: Camille Lamy <clamy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1498688}