Reland "[3PAutofillSettings] Change the disable settings explanation to a card"
This is a reland of commit f32305ad58c806d188d856776d343ed592cb6d4f

The fix for the failing test was to go back in the end of
testDisabledSettingsText_linksToAutofillOptionsPage so that the options
fragment that was opened in that test gets removed. The tests were
leaking state so another test would fail when the options fragment
remained open. The tests are now unbatched for safety
(crrev.com/c/6827804).

If the tests start to fail again please rather disable them then revert
because of the strings introduced in the CL. They will be hard to merge
back.

Original change's description:
> [3PAutofillSettings] Change the disable settings explanation to a card
>
> This changes the explanation why the Chrome settings are disabled from a
> simple text with a link to a card with a button.
>
> Screenshots: please look at the attached screenshots in the bug.
>
> Bug: 432445776
> Change-Id: I9d2b2a26110f6fbbb0cdfb9a67201098ce18e8e3
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6769781
> Reviewed-by: Christoph Schwering <schwering@google.com>
> Commit-Queue: Ivana Žužić <izuzic@google.com>
> Reviewed-by: Jan Keitel <jkeitel@google.com>
> Reviewed-by: Timofey Chudakov <tchudakov@google.com>
> Reviewed-by: Side YILMAZ <sideyilmaz@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1496492}

Bug: 432445776
Change-Id: Ib4c817f3c6a440a8315ba3c0df28389f9b7f4991
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6827143
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Cr-Commit-Position: refs/heads/main@{#1498711}