Allow the web client to select credentials
This CL introduce a way for the execution engine to prompt a user to
select a credential among many.

The the execution engine is not wired to the glic API yet, thus no
behavior change. The wiring will be done in a follow up, plus testing
via the test WebClient.

Bug: b:427817882
Change-Id: I89cdeacd84f4bb93db47dd882b39d0285338419a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6829002
Auto-Submit: William Liu <liuwilliam@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1498778}