[APC] Stop the flow on cross origin navigation
Currently when OTP is detected, we postpone the flow and wait until the
next page loads. There is a bug that if user doesn't fill the OTP and
navigates to a different page instead, we still start the flow. The
issue is fixed in the current CL by creating the
CrossOriginNavigationObserver, which would fire
OnCrossOriginNavigationDetected and stop the flow.

Bug: 435179350
Change-Id: I160144a3f0c0663802ec48967d9a0747b496e13e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6825943
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Cr-Commit-Position: refs/heads/main@{#1498861}