[Extensions] Remove allow_inline_enum code in schema compiler
Inline enum support was removed a long time ago (crbug.com/40411915) and
the only instances of it remaining were in some test files. This CL
fully removes support for them from the schema compiler and updates the
related test files with the current versions of the related APIs they
were copied from. This also required a few updates to related tests, to
reflect small changes that have happened to the APIs since then, as well
as to add the Enums into the list of expected Types exposed on the APIs.

Bug: N/A
Change-Id: Ie7c78dc1f3dc1faaf2ff1ab982a08a09936c9518
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6828744
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Tim <tjudkins@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1498878}