cc::SoftwareImageDecode: Cleanup before adding gainmap support
Rename the accessor cc::PaintImage::HasGainmap to
cc::PaintImage::HasGainmapInfo, to reflect that all that can be known is
that the gainmap metadata is present (the decode of the gainmap image
itself can fail).

Remove the helper function cc::ToneMapUtil::UseGainmapShader, because
the HasGainmap accessor is more directly useful (when we're about to use
the gainmap shader, we need to know if the gainmap was successfully
decoded and have it in hand).

Correctly propagate the target HDR headroom to the playback parameters
in RasterSource::PlaybackDisplayListToCanvas.

Bug: 428575083
Change-Id: I78adb03767c3efc9c070aa50b04b780436f24479
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6825116
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1498895}