SoftwareImageDecodeCache: Clean up CacheEntry
This is to prepare for adding gainmap support.

Change AllocateDiscardable to also populate a raster SkImage. This is
good for safety on its own (moving the allocation closer to where it
is used). When gain maps are added, they will be using the same
DiscardableMemory allocation for both images. That will require
pointer arithmetic on the DiscardableMemory. Introduce a base::span
representation of the DiscardableMemory.

In CacheEntry, we store an SkImageInfo but never actually use it
outside of the constructor, which uses it to create an SkImage. By
moving the SkImage creation to AllocateDiscardable, we allow removing
this memory. This will make things less cluttered when we add another
SkImage for the gainmap image.

Slightly clean up some of the surrounding code so that the gainmap
implementation will sit more nicely.

Bug: 428575083
Change-Id: If77c9bb7f0465050a598a6b7b5de1e58a3b0d670
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6828892
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1498896}