[PDF Ink Signatures] Change GetSelectionRects() to GetSelectionRectMap()
Currently PDFiumEngine::GetSelectionRects() returns the selection rects,
and the caller has to map each rect to the page it's on. However,
PDFiumEngine already knows the mapping, so the mapping work in the
GetSelectionRects() caller is actually unnecessary.

Instead, change GetSelectionRects() to GetSelectionRectMap() and return
a map of vectors, indexed by page, which is very to do inside
PDFiumEngine. Now the caller gets a ready-to-use map. Then update tests
that interact with GetSelectionRects() so they match the new behavior.

Change-Id: I0decd321d52f7fc049c8bbf0b16c519250d80ee2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6825315
Reviewed-by: Andy Phan <andyphan@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1498948}