[gap-decorations] Interpolation with repeaters for *-rule-color
This CL implements Interpolation for *-rule-color for repeaters.

In order to handle repeaters, we add a new InterpolableType
`InterpolableGapDataRepeater` which handles interpolation for repeaters
by essentially keeping a member InterpolableList of the values in the
repeater and interpolating those individually through that type.

The overall structure of interpolation for gap decorations will be as
follows:

`CSSGapLengthListInterpolationType` handles interpolation for
`GapDataList` We then hand off the actual interpolation with an
`InterpolableList` of `InterpolableLength`,
`InterpolableGapDataRepeater` (which itself contains an
`InterpolableList` of `InterpolableLengths`) and `InterpolableColor`.

InterpolableGapDataRepeater is templated, in order to handle for
interpolation for column/row-rule-color as well as width.

For now we are implementing mirroring the behavior of `grid-template`
interpolation (which can be a list of values with repeaters), in which
we only interpolate if the length of the lists match. If this changes
after spec discussions then the test will be updated.

https://www.w3.org/TR/css-gaps-1/#lists-repeat

Bug: 357648037, 419066541

Change-Id: I317b0c059a901d8eafa722aa71b878c4ecde3eda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6734136
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1498969}