[PWA/Trusted Icons] Read cached data from web app for icon size, metrics
This CL does 2 things:

1. Adds a new WebAppRegistrar API that reads the cached size data from
the web app to realize if there are stored trusted icons on the disk,
and falling back to manifest icons if needed. This is required for
callsites which pass in size as an input to the WebAppIconManager to read app icons, and so far, there was a disparity where the size information was obtained from the cached manifest icon downloaded sizes, but the icons being read were trusted ones. This CL fixes that disparity and updates existing tests.

2. Adds metrics to measure number of apps having trusted icons vs apps
having no trusted icons, along with a basic unit-test. This is useful to measure how often trusted icons are being populated, as the latter metric is expected to come down to zero sometime in the future.

Bug: 427566193
Change-Id: I1dd7b630f8746281e491933c14e1495ee0ef3ad5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6830111
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1498976}