Add NTPShortcuts policy YAML
Add policy yaml definition for NTPShortcuts, which will allow admins
to set shortcuts on the New Tab page.

This change does not include policy to pref mapping as a later change
will add that. This change is to unblock development work for the
Admin Console.

Bug: b:437359067
Change-Id: Ifd8820add8c93a82009385bef2cf80b7d3dab596
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6832712
Reviewed-by: Artem Sumaneev <asumaneev@google.com>
Reviewed-by: Sébastien Lalancette <seblalancette@chromium.org>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Alex Chen <alexwchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1498979}