[FedCM] Do not collect training data when UI is not shown
We only want to collect data when UI has been shown. This is especially
important for the case where the UI is suppressed by the model itself.
To do this, we need to move the did_show_ui_ to the controller, where
it better belongs. This CL does not change the fact we do not collect
training data when we do not even attempt to show UI, such as during
cooldown.

Bug: 435230692
Change-Id: I81fc077e17ecd7724af89a9533e712e256b2d312
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6832514
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Nicolás Peña <npm@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Yi Gu <yigu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1499013}