[Media] Fix Symphonia fuzzed CHECK crash
This patch fixes a bug in the Symphonia implementation found by
fuzzing, and caused by Symphonia returning an empty buffer when
there is an "Ok" status return. Previous Symphonia documentation
indicated that Symphonia returns an EndOfStream Error whenever the
end of the stream is met, however further investigation indicates
that, if the end of stream is expected, an OK status with an empty
buffer may be returned instead.

This patch adds a C++-side CHECK to make sure that the buffer is not marked
end of stream if the status is OK, and a Rust side change to ensure that
we return an EndOfStream status if the buffer is empty.

Bug: 433558757
Change-Id: Id41adcc3f6be4e7b8418dc8ff892c2ce0c72f84a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6822380
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1499053}