[iOS] Increase timeout for waitUntilReaderModeWebStateIsReady
At the moment there is flakiness in tests which depend on the result of
waitUntilReaderModeWebStateIsReady. The reason is the current timeout is
only 6 seconds (default for JS completion), whereas the Reader mode
WebState also needs to load the content to be ready. This CL updates
waitUntilReaderModeWebStateIsReady so it uses the sum
kWaitForJSCompletionTimeout + kWaitForPageLoadTimeout as timeout.

Bug: 436842225
Change-Id: Ib00dbbbe3ebd0621b755ca23e81d4ffa30dfff5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6829982
Commit-Queue: Quentin Pubert <qpubert@google.com>
Owners-Override: Asami Doi <asamidoi@chromium.org>
Auto-Submit: Quentin Pubert <qpubert@google.com>
Reviewed-by: Asami Doi <asamidoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1499450}