[RSG] Introduce ComposeSuggestionGenerator
This CL introduces a SuggestionGenerator implementation for compose.
Compared to the other suggestion generators, compose does not rely on
any suggestion data in order to produce suggestions, so
FetchSuggestionData() returns always with an empty vector and the
GenerateSuggestions() method does the heavy lifting of checking for
the field requirements.

Bug: 409962888
Change-Id: I090def184489c7f374c69e3406c9aef3705b9f3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6827221
Commit-Queue: Karol Sygiet <sygiet@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1499452}