Persist the latest variations seed serial number in a preference.
Instead of loading from the seed and caching the latest serial number in
a member variable, store it in a dedicated Local State preference.

The serial number will be persisted between chrome sessions without need
to load the seed on startup. Since we're trying to remove the seed from
memory when it's not used, this will also be beneficial in the future,
since it will avoid an extra read-file operation.

Note: Since this is a new pref, the pref will need to be populated
before it will have a value. For most clients, this will happen via the
code added to VariationsSeedStore::LoadSeed(). However, for clients
using a safe seed, that function won't be called and we'll return an
empty string instead. This will result in the clients not sending the
serial number to the variations server, and fetching a full seed instead
of a delta-compressed one. This will be a one-time effect, as after the
seed is fetched, we'll populate the pref and its value will be available
from then on.

Bug: 436429062
Change-Id: I2c82241181f910355646c9d4a81c38c3fe6a93eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6818815
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Ramon Cano Aparicio <rcanoaparicio@google.com>
Cr-Commit-Position: refs/heads/main@{#1499482}