[WebAuthn] Simplify WebAuthn JNI credential callbacks
This is a follow-up from
https://chromium-review.googlesource.com/c/chromium/src/+/6794224.

The `onCredentialsDetailsListReceived` JNI method in
`WebauthnBrowserBridge` has become overly complicated. This is a
refactoring CL to simplify it.

It combines password and passkey credentials returned from native into
a single callback, and that callback is no longer used to return error
states as the passkey one was previously.

Instead, the callback for closing immediate UI has been generalized
to cover any non-credential completion, and renamed for that purpose.

The hybrid callback has been left as is because its state of being
null or non-null informs the UI whether to offer the button.

Bug: 433543129
Change-Id: I3e96835611d29c55857b52c856af1ca170721edc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6829407
Reviewed-by: Adem Derinel <derinel@google.com>
Reviewed-by: Friedrich Hauser <friedrichh@chromium.org>
Commit-Queue: Friedrich Hauser <friedrichh@chromium.org>
Auto-Submit: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1499515}