[ios] Do not retrieve grey snapshots behind the flag
This CL adds a feature flag and prevent from retrieving grey snapshots
when the flag is enabled. A tab is blank while loading a page (see the
screen recording below). The code for grey snapshots will be deleted
once the metrics look ok.

The feature is enabled:
https://drive.google.com/file/d/1IhYkNn72IGUGuMBey1TXZECvUOKnFB4b/view?usp=sharing
The feature is disabled:
https://drive.google.com/file/d/1uV619iN3cOkPlr_4zEm5kolB4L6Kj9Uy/view?usp=sharing

Bug: 433236646
Change-Id: I5f88ab849b5d70460808b28669132a641ea5dabb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6828865
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1499536}