[Extensions] Track callbacks for one time messages with a unique token
Before this change we tracked pending response callbacks for one time
messages by using the memory address of the callback. This wasn't an
issue before the promise support feature because we didn't allocate
callbacks into the OneTimeMessageContextData's pending_callbacks vector
asynchronously.

With the promise support feature we do allocate additional
(promise-related) callbacks into the pending_callbacks vector
asynchronously after the response callback is allocated. This can cause
the promise callback to be called if the extension script calls the
response callback multiple times. For example (in chronological order):

1. A response callback is stored in a list called pending_callbacks at
  memory address M1.
2. The extension script runs this callback, then the callback is removed
  from the list.
3. By coincidence, a new promise callback is created and allocated the
  same memory address (M1), and is then stored in the list.
4. The script context can still be alive and the extension script then
  tries to send another response, attempting to call the callback at
  memory address M1.
5. Instead of finding the original response callback, it finds and
  incorrectly runs the new promise callback.

After this change pending_callbacks are now tracked with an
base::UnguessableToken which should be unique enough to not reasonably
allow this reuse of memory addresses/CallbackIDs to happen.

This now allows listeners to return promises that reject without this
possible bug from occurring.

Co-authored-by: rdevlin.cronin@chromium.org
Bug: 433214074
Change-Id: Ia7096f0200e6bc8617b9ca1b25227cc20f37887b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6819178
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Justin Lulejian <jlulejian@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1499596}